+++ src/run_error.rs
@@ -1,8 +1,8 @@
 use crate::parse_error::OMLCodeReason;
 use crate::{ConfReason, config_error::ConfCore};
 use derive_more::From;
-use orion_error::reason::ConfErrReason;
-use orion_error::{OrionError, StructError, UvsFrom, UvsReason};
+
+use orion_error::{OrionError, StructError, UnifiedReason};
 use orion_sec::OrionSecReason;
 use serde::Serialize;
 use thiserror::Error;
@@ -39,25 +39,25 @@ pub enum RunReason {
     #[orion_error(identity = "biz.source")]
     Source(SourceFocus),
     #[orion_error(transparent)]
-    Uvs(UvsReason),
+    Uvs(UnifiedReason),
 }
 
 impl From<ConfReason<ConfCore>> for RunReason {
     fn from(_: ConfReason<ConfCore>) -> Self {
-        Self::from_conf_reason(ConfErrReason::Core)
+        Self::core_conf()
     }
 }
 impl From<OMLCodeReason> for RunReason {
     fn from(_: OMLCodeReason) -> Self {
-        Self::from_conf()
+        Self::core_conf()
     }
 }
 
 impl From<OrionSecReason> for RunReason {
     fn from(value: OrionSecReason) -> Self {
         match value {
-            OrionSecReason::Sec(_) => Self::from_conf(),
-            OrionSecReason::Uvs(uvs_reason) => Self::Uvs(uvs_reason),
+            OrionSecReason::Sec(_) => Self::core_conf(),
+            OrionSecReason::General(uvs_reason) => Self::Uvs(uvs_reason),
         }
     }
 }
@@ -74,20 +74,18 @@ where
     fn owe_sink(self) -> RunResult<T> {
         match self {
             Ok(v) => Ok(v),
-            Err(e) => Err(StructError::from(RunReason::Dist(DistFocus::SinkError(
-                "sink fail".into(),
-            )))
-            .with_detail(e.to_string())),
+            Err(e) => Err(RunReason::Dist(DistFocus::SinkError("sink fail".into()))
+                .to_err()
+                .with_detail(e.to_string())),
         }
     }
     fn owe_source(self) -> RunResult<T> {
         match self {
             Ok(v) => Ok(v),
             Err(e) => Err(
-                StructError::from(RunReason::Source(SourceFocus::SupplierError(
-                    "source fail".into(),
-                )))
-                .with_detail(e.to_string()),
+                RunReason::Source(SourceFocus::SupplierError("source fail".into()))
+                    .to_err()
+                    .with_detail(e.to_string()),
             ),
         }
     }
@@ -131,14 +129,15 @@ impl From<SinkReason> for RunReason {
         }
     }
 }
+use orion_conf::ToStructError as _;
 use orion_conf::error::ConfIOReason;
 use wp_connector_api::{SinkReason, SourceReason};
 impl From<ConfIOReason> for RunReason {
     fn from(value: ConfIOReason) -> Self {
         match value {
-            ConfIOReason::Other(_) => RunReason::from_conf(),
-            ConfIOReason::Uvs(uvs) => RunReason::Uvs(uvs),
-            ConfIOReason::NoFormatEnabled => RunReason::from_conf(),
+            ConfIOReason::Other(_) => RunReason::core_conf(),
+            ConfIOReason::General(uvs) => RunReason::Uvs(uvs),
+            ConfIOReason::NoFormatEnabled => RunReason::core_conf(),
         }
     }
 }